feat: the function to convert from roman numerals to arabics is completed
The new funtion romanToArabic has been added into the utility class. Tto
create this function, some basic rules about Roman Numerals have been
implemented, this information is going to be added into the README.
Basically, the received string is going to be iterated and depending on
the next symbol, the corresponfing values is added or subtracted from
the total sum.